Pdev 1.9.3

+++ game/Submods/Dialogue Packs/P/pc_update_helper.rpy
@@ -33,7 +33,8 @@ init python:
         #复制文件
         cDir = renpy.config.basedir + "/" + dir
         tDir = subUBasedir + "/" + dir
-        shutil.copytree(cDir,tDir)
+        if os.path.exists(cDir):
+            shutil.copytree(cDir,tDir)
         return
 
 init 5 python:
@@ -50,11 +51,11 @@ init 5 python:
         )
 
 label sub_update_helper:
-    m "好的"
-    m "如果未响应, 这是正常情况, 因为在干活要稍微等一下..."
+    m 1eua "好的"
+    m 3eka "如果未响应, 这是正常情况, 因为在干活要稍微等一下..."
     call copyfile
-    m "搞定了. {w=0.3}文件都在character/OldVersionFiles文件夹里, 复制到新版本就好了."
-    m "不要复制到新版本的character了, 要去合并新版本game文件夹."
+    m 4eub "搞定了. {w=0.3}文件都在character/OldVersionFiles文件夹里, 复制到新版本就好了."
+    m 3eua "不要复制到新版本的character了, 要去合并新版本game文件夹."
 return
 
 label copyfile:
@@ -67,7 +68,7 @@ label copyfile:
         createFileBaseDir()
         doingPerc = 0
         for lista in todoDir:
-            renpy.say(m,"[todoDir.index(lista)+1]/[len(todoDir)] - 正在复制文件夹:"+lista+"{fast}{nw}")
+            renpy.say(m,"[todoDir.index(lista)+1]/[len(todoDir)] | 正在复制文件夹:"+lista+"{fast}{nw}")
             copyOldDir(lista)
             
     return
